python-backend/requirements.txt
openai-agents
openai-chatkit
pydantic
flask
